requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pydantic
pandas
plotly
python-dotenv
